Revert to using templates and clean up
