+++ SDL/LST.cc
@@ -9,109 +9,6 @@
 #include "Math/Vector3D.h"
 using XYZVector = ROOT::Math::XYZVector;
 
-namespace {
-  std::string trackLooperDir() {
-    const char* path_lst_base = std::getenv("LST_BASE");
-    const char* path_tracklooperdir = std::getenv("TRACKLOOPERDIR");
-    std::string path_str;
-    if (path_lst_base != nullptr) {
-      path_str = path_lst_base;
-    } else if (path_tracklooperdir != nullptr) {
-      path_str = path_tracklooperdir;
-    } else {
-      // FIXME: temporary solution, will need to pass a value from FileInPath or CMSSW search path
-      // in the `LSTProducer` or a related ES producer
-      path_str = std::getenv("CMSSW_BASE");
-      path_str += "/src/RecoTracker/LSTCore/TrackLooper";
-    }
-    return path_str;
-  }
-
-  std::string get_absolute_path_after_check_file_exists(const std::string name) {
-    std::filesystem::path fullpath = std::filesystem::absolute(name.c_str());
-    if (not std::filesystem::exists(fullpath)) {
-      std::cout << "ERROR: Could not find the file = " << fullpath << std::endl;
-      exit(2);
-    }
-    return fullpath.string();
-  }
-
-  void loadMapsHost(SDL::MapPLStoLayer& pLStoLayer,
-                    std::shared_ptr<SDL::EndcapGeometry<SDL::DevHost, false>> endcapGeometry,
-                    std::shared_ptr<SDL::TiltedGeometry> tiltedGeometry,
-                    std::shared_ptr<SDL::ModuleConnectionMap> moduleConnectionMap) {
-    // Module orientation information (DrDz or phi angles)
-    auto endcap_geom =
-        get_absolute_path_after_check_file_exists(trackLooperDir() + "/data/OT800_IT615_pt0.8/endcap_orientation.bin");
-    auto tilted_geom = get_absolute_path_after_check_file_exists(
-        trackLooperDir() + "/data/OT800_IT615_pt0.8/tilted_barrel_orientation.bin");
-    // Module connection map (for line segment building)
-    auto mappath = get_absolute_path_after_check_file_exists(
-        trackLooperDir() + "/data/OT800_IT615_pt0.8/module_connection_tracing_merged.bin");
-
-    endcapGeometry->load(endcap_geom);
-    tiltedGeometry->load(tilted_geom);
-    moduleConnectionMap->load(mappath);
-
-    auto pLSMapDir = trackLooperDir() + "/data/OT800_IT615_pt0.8/pixelmap/pLS_map";
-    const std::array<std::string, 4> connects{
-        {"_layer1_subdet5", "_layer2_subdet5", "_layer1_subdet4", "_layer2_subdet4"}};
-    std::string path;
-
-    static_assert(connects.size() == std::tuple_size<std::decay_t<decltype(pLStoLayer[0])>>{});
-    for (unsigned int i = 0; i < connects.size(); i++) {
-      auto connectData = connects[i].data();
-
-      path = pLSMapDir + connectData + ".bin";
-      pLStoLayer[0][i] = SDL::ModuleConnectionMap(get_absolute_path_after_check_file_exists(path));
-
-      path = pLSMapDir + "_pos" + connectData + ".bin";
-      pLStoLayer[1][i] = SDL::ModuleConnectionMap(get_absolute_path_after_check_file_exists(path));
-
-      path = pLSMapDir + "_neg" + connectData + ".bin";
-      pLStoLayer[2][i] = SDL::ModuleConnectionMap(get_absolute_path_after_check_file_exists(path));
-    }
-  }
-
-}  // namespace
-
-std::unique_ptr<SDL::LSTESHostData> SDL::loadAndFillESHost() {
-  auto pLStoLayer = std::make_shared<SDL::MapPLStoLayer>();
-  auto endcapGeometry = std::make_shared<SDL::EndcapGeometry<SDL::DevHost, false>>();
-  auto tiltedGeometry = std::make_shared<SDL::TiltedGeometry>();
-  auto moduleConnectionMap = std::make_shared<SDL::ModuleConnectionMap>();
-  ::loadMapsHost(*pLStoLayer, endcapGeometry, tiltedGeometry, moduleConnectionMap);
-  return std::make_unique<LSTESHostData>(pLStoLayer, endcapGeometry, tiltedGeometry, moduleConnectionMap);
-}
-
-std::unique_ptr<SDL::LSTESDeviceData<SDL::Dev>> SDL::loadAndFillESDevice(SDL::QueueAcc& queue,
-                                                                         const LSTESHostData* hostData) {
-  SDL::Dev const& devAccIn = alpaka::getDev(queue);
-  uint16_t nModules;
-  uint16_t nLowerModules;
-  unsigned int nPixels;
-  std::shared_ptr<SDL::modulesBuffer<SDL::Dev>> modulesBuffers = nullptr;
-  auto endcapGeometry =
-      std::make_shared<SDL::EndcapGeometry<SDL::Dev, true>>(devAccIn, queue, *hostData->endcapGeometry);
-  auto pixelMapping = std::make_shared<SDL::pixelMap>();
-  auto moduleConnectionMap = hostData->moduleConnectionMap;
-
-  auto path =
-      get_absolute_path_after_check_file_exists(trackLooperDir() + "/data/OT800_IT615_pt0.8/sensor_centroids.bin");
-  SDL::loadModulesFromFile(queue,
-                           hostData->mapPLStoLayer.get(),
-                           path.c_str(),
-                           nModules,
-                           nLowerModules,
-                           nPixels,
-                           modulesBuffers,
-                           pixelMapping.get(),
-                           endcapGeometry.get(),
-                           hostData->tiltedGeometry.get(),
-                           moduleConnectionMap.get());
-  return std::make_unique<LSTESDeviceData<SDL::Dev>>(
-      nModules, nLowerModules, nPixels, modulesBuffers, endcapGeometry, pixelMapping);
-}
 
 void SDL::LST<SDL::Acc>::run(SDL::QueueAcc& queue,
                              bool verbose,
