+++ SDL/ModuleConnectionMap.h
@@ -15,7 +15,11 @@
 #endif
 
 namespace SDL {
-  class ModuleConnectionMap {
+  //FIXME: move to non-alpaka single arch build
+  template <typename>
+  class ModuleConnectionMap;
+  template <>
+  class ModuleConnectionMap<SDL::Dev> {
   private:
     std::map<unsigned int, std::vector<unsigned int>> moduleConnections_;
 
@@ -32,7 +36,7 @@ namespace SDL {
     int size() const;
   };
 
-  using MapPLStoLayer = std::array<std::array<ModuleConnectionMap, 4>, 3>;
+  using MapPLStoLayer = std::array<std::array<ModuleConnectionMap<SDL::Dev>, 4>, 3>;
 }  // namespace SDL
 
 #endif
