+++ SDL/TiltedGeometry.cc
@@ -1,8 +1,8 @@
 #include "TiltedGeometry.h"
 
-SDL::TiltedGeometry::TiltedGeometry(std::string filename) { load(filename); }
+SDL::TiltedGeometry<SDL::Dev>::TiltedGeometry(std::string filename) { load(filename); }
 
-void SDL::TiltedGeometry::load(std::string filename) {
+void SDL::TiltedGeometry<SDL::Dev>::load(std::string filename) {
   drdzs_.clear();
   dxdys_.clear();
 
@@ -32,7 +32,7 @@ void SDL::TiltedGeometry::load(std::string filename) {
   }
 }
 
-float SDL::TiltedGeometry::getDrDz(unsigned int detid) const {
+float SDL::TiltedGeometry<SDL::Dev>::getDrDz(unsigned int detid) const {
   if (drdzs_.find(detid) != drdzs_.end()) {
     return drdzs_.at(detid);
   } else {
@@ -40,7 +40,7 @@ float SDL::TiltedGeometry::getDrDz(unsigned int detid) const {
   }
 }
 
-float SDL::TiltedGeometry::getDxDy(unsigned int detid) const {
+float SDL::TiltedGeometry<SDL::Dev>::getDxDy(unsigned int detid) const {
   if (dxdys_.find(detid) != dxdys_.end()) {
     return dxdys_.at(detid);
   } else {
