+++ code/core/AnalysisConfig.h
@@ -103,7 +103,7 @@ public:
     std::map<int, std::vector<std::vector<float>>> moduleSimHits;
     std::map<int, int> modulePopulation;
 
-    SDL::ModuleConnectionMap moduleConnectiongMapLoose;
+    SDL::ModuleConnectionMap<SDL::Dev> moduleConnectiongMapLoose;
 
     // Boolean to trigger whether to run cut_value_ntupling
     bool do_cut_value_ntuple;
